# triedb/pathdb: introduce lookup structure to optimize node query