--- a/core/state/trie_prefetcher.go
+++ b/core/state/trie_prefetcher.go
@@ -19,6 +19,7 @@ package state
 import (
 	"errors"
 	"sync"
+	"time"
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/log"
@@ -55,6 +56,7 @@ type triePrefetcher struct {
 	accountDupWriteMeter  *metrics.Meter
 	accountDupCrossMeter  *metrics.Meter
 	accountWasteMeter     *metrics.Meter
+	accountLoadTimer      *metrics.ResettingTimer
 
 	storageLoadReadMeter  *metrics.Meter
 	storageLoadWriteMeter *metrics.Meter
@@ -62,6 +64,7 @@ type triePrefetcher struct {
 	storageDupWriteMeter  *metrics.Meter
 	storageDupCrossMeter  *metrics.Meter
 	storageWasteMeter     *metrics.Meter
+	storageLoadTimer      *metrics.ResettingTimer
 }
 
 func newTriePrefetcher(db Database, root common.Hash, namespace string, noreads bool) *triePrefetcher {
@@ -78,6 +81,7 @@ func newTriePrefetcher(db Database, root common.Hash, namespace string, noreads 
 
 		accountLoadReadMeter:  metrics.GetOrRegisterMeter(prefix+"/account/load/read", nil),
 		accountLoadWriteMeter: metrics.GetOrRegisterMeter(prefix+"/account/load/write", nil),
+		accountLoadTimer:      metrics.GetOrRegisterResettingTimer(prefix+"/account/load/time", nil),
 		accountDupReadMeter:   metrics.GetOrRegisterMeter(prefix+"/account/dup/read", nil),
 		accountDupWriteMeter:  metrics.GetOrRegisterMeter(prefix+"/account/dup/write", nil),
 		accountDupCrossMeter:  metrics.GetOrRegisterMeter(prefix+"/account/dup/cross", nil),
@@ -85,6 +89,7 @@ func newTriePrefetcher(db Database, root common.Hash, namespace string, noreads 
 
 		storageLoadReadMeter:  metrics.GetOrRegisterMeter(prefix+"/storage/load/read", nil),
 		storageLoadWriteMeter: metrics.GetOrRegisterMeter(prefix+"/storage/load/write", nil),
+		storageLoadTimer:      metrics.GetOrRegisterResettingTimer(prefix+"/storage/load/time", nil),
 		storageDupReadMeter:   metrics.GetOrRegisterMeter(prefix+"/storage/dup/read", nil),
 		storageDupWriteMeter:  metrics.GetOrRegisterMeter(prefix+"/storage/dup/write", nil),
 		storageDupCrossMeter:  metrics.GetOrRegisterMeter(prefix+"/storage/dup/cross", nil),
@@ -121,6 +126,10 @@ func (p *triePrefetcher) report() {
 			p.accountLoadReadMeter.Mark(int64(len(fetcher.seenReadAddr)))
 			p.accountLoadWriteMeter.Mark(int64(len(fetcher.seenWriteAddr)))
 
+			total := len(fetcher.seenReadAddr) + len(fetcher.seenWriteAddr)
+			if total > 0 {
+				p.accountLoadTimer.Update(fetcher.readTime / time.Duration(total))
+			}
 			p.accountDupReadMeter.Mark(int64(fetcher.dupsRead))
 			p.accountDupWriteMeter.Mark(int64(fetcher.dupsWrite))
 			p.accountDupCrossMeter.Mark(int64(fetcher.dupsCross))
@@ -134,6 +143,10 @@ func (p *triePrefetcher) report() {
 			p.storageLoadReadMeter.Mark(int64(len(fetcher.seenReadSlot)))
 			p.storageLoadWriteMeter.Mark(int64(len(fetcher.seenWriteSlot)))
 
+			total := len(fetcher.seenReadSlot) + len(fetcher.seenWriteSlot)
+			if total > 0 {
+				p.storageLoadTimer.Update(fetcher.readTime / time.Duration(total))
+			}
 			p.storageDupReadMeter.Mark(int64(fetcher.dupsRead))
 			p.storageDupWriteMeter.Mark(int64(fetcher.dupsWrite))
 			p.storageDupCrossMeter.Mark(int64(fetcher.dupsCross))
@@ -241,6 +254,7 @@ type subfetcher struct {
 	seenWriteAddr map[common.Address]struct{} // Tracks the accounts already loaded via write operations
 	seenReadSlot  map[common.Hash]struct{}    // Tracks the storage already loaded via read operations
 	seenWriteSlot map[common.Hash]struct{}    // Tracks the storage already loaded via write operations
+	readTime      time.Duration               // Total time spent on state resolving
 
 	dupsRead  int // Number of duplicate preload tasks via reads only
 	dupsWrite int // Number of duplicate preload tasks via writes only
@@ -388,6 +402,7 @@ func (sf *subfetcher) loop() {
 			sf.tasks = nil
 			sf.lock.Unlock()
 
+			start := time.Now()
 			for _, task := range tasks {
 				if task.addr != nil {
 					key := *task.addr
@@ -451,6 +466,10 @@ func (sf *subfetcher) loop() {
 					}
 				}
 			}
+			// Count the time being spent on state resolving. While it's not very
+			// accurate due to some additional operations (e.g., filter out duplicated
+			// task), but it's already good enough for monitoring.
+			sf.readTime += time.Since(start)
 
 		case <-sf.stop:
 			// Termination is requested, abort if no more tasks are pending. If