--- a/triedb/pathdb/reader.go
+++ b/triedb/pathdb/reader.go
@@ -17,6 +17,7 @@
 package pathdb
 
 import (
+	"errors"
 	"fmt"
 
 	"github.com/ethereum/go-ethereum/common"
@@ -50,15 +51,27 @@ func (loc *nodeLoc) string() string {
 // reader implements the database.NodeReader interface, providing the functionalities to
 // retrieve trie nodes by wrapping the internal state layer.
 type reader struct {
-	layer       layer
+	db          *Database
+	state       common.Hash
 	noHashCheck bool
+	layer       layer
 }
 
 // Node implements database.NodeReader interface, retrieving the node with specified
 // node info. Don't modify the returned byte slice since it's not deep-copied
 // and still be referenced by database.
 func (r *reader) Node(owner common.Hash, path []byte, hash common.Hash) ([]byte, error) {
-	blob, got, loc, err := r.layer.node(owner, path, 0)
+	l := r.db.tree.lookupNode(owner, path, r.state)
+	if l == nil {
+		return nil, errors.New("node is not found")
+	}
+	// Check staleness after querying the lookup set. Otherwise, there is a
+	// theoretical possibility that the layer could be marked as stale after
+	// the initial staleness check.
+	if r.layer.isStale() {
+		return nil, errSnapshotStale
+	}
+	blob, got, loc, err := l.node(owner, path, 0)
 	if err != nil {
 		return nil, err
 	}
@@ -136,7 +149,12 @@ func (db *Database) NodeReader(root common.Hash) (database.NodeReader, error) {
 	if layer == nil {
 		return nil, fmt.Errorf("state %#x is not available", root)
 	}
-	return &reader{layer: layer, noHashCheck: db.isVerkle}, nil
+	return &reader{
+		db:          db,
+		state:       root,
+		noHashCheck: db.isVerkle,
+		layer:       layer,
+	}, nil
 }
 
 // StateReader returns a reader that allows access to the state data associated