Implemented the https://github.com/ethereum/go-ethereum/issues/30022

Benchmark needed