@rjl493456442 2024-12-04
After running the benchmark for a few days, these are the conclusion I get:

> Bench07 is the pull request; Bench08 is the master;

(a) the performance of node retrieval is improved significantly

The chart illustrates the time required for account trie updates. Clearly, bench07 is significantly faster than master.
More importantly, the prefetcher resolves trie nodes in advance at the transaction boundary, with most nodes being processed in the background. As a result, the actual performance difference is even greater than what the chart shows.

<img width="1445" alt="截屏2024-12-04 21 00 18" src="https://github.com/user-attachments/assets/4692c71f-6f2d-4013-8221-addcdca92cc6">

What's more, the difference can also be observed from the CPU profiling. As we can see about 8.56% cpu time is spent on map access which is caused by pathdb layer iteration. These two parts have been eliminated in PR.

```
[ Master ]
Showing nodes accounting for 247.19s, 42.68% of 579.16s total
Dropped 2211 nodes (cum <= 2.90s)
Showing top 10 nodes out of 314
      flat  flat%   sum%        cum   cum%
    44.53s  7.69%  7.69%     52.65s  9.09%  runtime.findObject
    41.84s  7.22% 14.91%     42.53s  7.34%  runtime.cgocall                               (ecrecover)
    31.51s  5.44% 20.35%     38.55s  6.66%  runtime.mapaccess2_faststr                    (difflayer.Node)
    28.81s  4.97% 25.33%     28.81s  4.97%  golang.org/x/crypto/sha3.keccakF1600
    25.94s  4.48% 29.81%    141.79s 24.48%  runtime.scanobject
    18.05s  3.12% 32.92%     26.66s  4.60%  runtime.mapaccess2                            (difflayer.Node)
    14.88s  2.57% 35.49%     14.88s  2.57%  internal/runtime/syscall.Syscall6
    14.50s  2.50% 38.00%     14.50s  2.50%  sync/atomic.(*Int32).Add
    13.66s  2.36% 40.35%     13.84s  2.39%  runtime.(*gcBits).bitp (inline)
    13.47s  2.33% 42.68%     49.84s  8.61%  runtime.mallocgc

[ PR ]
Showing nodes accounting for 208.28s, 42.26% of 492.80s total
Dropped 2301 nodes (cum <= 2.46s)
Showing top 10 nodes out of 337
      flat  flat%   sum%        cum   cum%
    44.87s  9.11%  9.11%     45.31s  9.19%  runtime.cgocall                              (ecrecover)                  
    41.51s  8.42% 17.53%     46.76s  9.49%  runtime.findObject
    29.38s  5.96% 23.49%     29.38s  5.96%  golang.org/x/crypto/sha3.keccakF1600
    18.48s  3.75% 27.24%    119.70s 24.29%  runtime.scanobject
    15.85s  3.22% 30.46%     15.85s  3.22%  internal/runtime/syscall.Syscall6
    14.35s  2.91% 33.37%     48.83s  9.91%  runtime.mallocgc
    13.94s  2.83% 36.20%     14.04s  2.85%  runtime.(*gcBits).bitp (inline)
    10.27s  2.08% 38.28%     10.43s  2.12%  runtime.(*mspan).heapBitsSmallForAddr
    10.20s  2.07% 40.35%     10.20s  2.07%  runtime.memmove
     9.43s  1.91% 42.26%     11.84s  2.40%  github.com/ethereum/go-ethereum/core/vm.codeBitmapInternal
```

(b) triedb commit is slower by having the additional lookup structure

This pull request introduces the lookup structure, which requires updates whenever a new block is processed. The performance difference observed here reflects the associated overhead.

Theoretically, this overhead can be optimized by parallelizing the lookup update procedure, which can save around 1-2ms per block. Unfortunately, the other parts of system (e.g. EVM execution) will be affected with unknown reason. Therefore the optimization is not included in this PR.

<img width="1440" alt="截屏2024-12-04 21 04 00" src="https://github.com/user-attachments/assets/651f4e2b-3a5c-4158-997e-9703bb8b5f12">

(c) otherwise, no major performance difference is observed

The PR is slightly faster than master. The main reason is for the ancient chain segments (before byzantium fork), prefetcher is basically useless and most of the nodes have to be resolved by the trie itself.

<img width="1442" alt="截屏2024-12-04 21 08 04" src="https://github.com/user-attachments/assets/1850c49d-c1b4-42a3-b2d4-4af6cff9fdb9">
<img width="1440" alt="截屏2024-12-04 21 08 55" src="https://github.com/user-attachments/assets/9fa04be2-4ce3-44c7-a10b-5371d02436ae">